Impl `ValidationCheck` for `HalfEdgeHasNoSibling`
