+++ crates/fj-core/src/validation/checks/half_edge_siblings.rs
@@ -1,4 +1,12 @@
-use crate::{storage::Handle, topology::HalfEdge};
+use std::collections::BTreeMap;
+
+use crate::{
+    geometry::Geometry,
+    queries::{BoundingVerticesOfHalfEdge, SiblingOfHalfEdge},
+    storage::Handle,
+    topology::{HalfEdge, Shell},
+    validation::{ValidationCheck, ValidationConfig},
+};
 
 /// A [`Shell`] contains a [`HalfEdge`] without a sibling
 ///
@@ -9,11 +17,58 @@ use crate::{storage::Handle, topology::HalfEdge};
 /// this is a sign of either of the following:
 /// - That the shell is not closed, meaning it has some kind of hole.
 /// - If the shell is closed, that its topological object graph is not valid.
-///
-/// [`Shell`]: crate::topology::Shell
 #[derive(Clone, Debug, thiserror::Error)]
 #[error("Half-edge has no sibling: {half_edge:#?}")]
 pub struct HalfEdgeHasNoSibling {
     /// The half-edge that does not have a sibling
     pub half_edge: Handle<HalfEdge>,
+}
+
+impl ValidationCheck<Shell> for HalfEdgeHasNoSibling {
+    fn check<'r>(
+        object: &'r Shell,
+        geometry: &'r Geometry,
+        _: &'r ValidationConfig,
+    ) -> impl Iterator<Item = Self> + 'r {
+        let mut unmatched_half_edges = BTreeMap::new();
+
+        for face in object.faces() {
+            for cycle in face.region().all_cycles() {
+                for half_edge in cycle.half_edges() {
+                    let curve = half_edge.curve().clone();
+                    let boundary = geometry.of_half_edge(half_edge).boundary;
+                    let vertices =
+                        cycle.bounding_vertices_of_half_edge(half_edge).expect(
+                            "`half_edge` came from `cycle`, must exist there",
+                        );
+
+                    let key = (curve.clone(), boundary, vertices.clone());
+                    let key_reversed =
+                        (curve, boundary.reverse(), vertices.reverse());
+
+                    match unmatched_half_edges.remove(&key_reversed) {
+                        Some(sibling) => {
+                            // This must be the sibling of the half-edge we're
+                            // currently looking at. Let's make sure the logic
+                            // we use here to determine that matches the
+                            // "official" definition.
+                            assert!(object
+                                .are_siblings(half_edge, sibling, geometry));
+                        }
+                        None => {
+                            // If this half-edge has a sibling, we haven't seen
+                            // it yet. Let's store this half-edge then, in case
+                            // we come across the sibling later.
+                            unmatched_half_edges.insert(key, half_edge);
+                        }
+                    }
+                }
+            }
+        }
+
+        unmatched_half_edges
+            .into_values()
+            .cloned()
+            .map(|half_edge| HalfEdgeHasNoSibling { half_edge })
+    }
 }
